requirements.txt
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
psycopg2==2.9.9
boto3==1.28.61
pandas==2.1.1
